refactor : package 구성 변경
