+++ src/main/kotlin/carracing/view/InputView.kt
@@ -1,4 +1,4 @@
-package carracing.ui
+package carracing.view
 
 import java.lang.IllegalArgumentException
